Add new methods related to newFolder naming
Add method to validate the name for a new folder
Add method to change the name of a new folder
Add method to check if a tree node is suitable as the installation
location.
Improve the compareName comparator so that the items will be sorted the
same way they are in the Finder.